More reorganization, docs, bump base version
