+++ core/shared/src/main/scala/org/polyvariant/colorize/package.scala
@@ -1,6 +1,5 @@
 package org.polyvariant
 
-package object colorize
-  extends org.polyvariant.colorize.ConfiguredColorize(
-    RenderConfig(RenderConfig.ColorMode.Ansi, Console.RESET)
-  )
+import org.polyvariant.colorize.custom._
+
+package object colorize extends ConfiguredColorize(RenderConfig.Default)
