+++ core/shared/src/main/scala/org/polyvariant/colorize/truecolor.scala
@@ -1,0 +1,6 @@
+package org.polyvariant.colorize
+
+import org.polyvariant.colorize.custom.ConfiguredColorize
+import org.polyvariant.colorize.custom.RenderConfig
+
+object trueColor extends ConfiguredColorize(RenderConfig.trueColor)
